Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,8 +1,8 @@
 #includ<stdio.h>
 int main()
 {
-    float a;
-    scanf("%f",&a);
-    printf("%.4f",a);
+    double a;
+    scanf("%lf",&a);
+    printf("%lf",a);
     return 0;
 }
